Refactor rootUid to rootUuid across sync engine and related files; update ServerFile type and remove unused FileCreatorFromServerFile
