+++ src/context/shared/domain/ServerFile.ts
@@ -11,7 +11,7 @@ export type ServerFile = {
   encrypt_version: string;
   fileId: string;
   folderId: number;
-  folderUuid?: string;
+  folderUuid: string;
   id: number;
   modificationTime: string;
   name: string;
