+++ src/context/virtual-drive/remoteTree/application/Traverser.ts
@@ -3,13 +3,13 @@ import { Service } from 'diod';
 import Logger from 'electron-log';
 import { ServerFile, ServerFileStatus } from '../../../shared/domain/ServerFile';
 import { ServerFolder, ServerFolderStatus } from '../../../shared/domain/ServerFolder';
-import { createFileFromServerFile } from './FileCreatorFromServerFile';
 import { createFolderFromServerFolder } from '../../folders/application/create/FolderCreatorFromServerFolder';
 import { Folder } from '../../folders/domain/Folder';
 import { FolderStatus, FolderStatuses } from '../../folders/domain/FolderStatus';
 import { EitherTransformer } from '../../shared/application/EitherTransformer';
 import { NameDecrypt } from '../domain/NameDecrypt';
 import { RemoteTree } from '../domain/RemoteTree';
+import { createFileFromServerFile } from '../../files/application/FileCreatorFromServerFile';
 
 type Items = {
   files: Array<ServerFile>;
